PK3: __len__ returned float instead of int. Typecasted.

+++ sklearn/cross_validation.py
@@ -170,7 +170,7 @@ class LeavePOut(object):
         )
 
     def __len__(self):
-        return (factorial(self.n) / factorial(self.n - self.p)
+        return int(factorial(self.n) / factorial(self.n - self.p)
                 / factorial(self.p))
 
 
@@ -528,7 +528,7 @@ class LeavePLabelOut(object):
         )
 
     def __len__(self):
-        return (factorial(self.n_unique_labels) /
+        return int(factorial(self.n_unique_labels) /
                 factorial(self.n_unique_labels - self.p) /
                 factorial(self.p))
